Change Option<> fields behavior
Currently there's a specialized behavior for Option<>
fields - they are being overridden even if the applied
field is none.

This is counter-intuitive, and doesn't make much sense (at least
for me).

Remove the specialized behavior and make it behave similarly
to any other field.

Note that this is a breaking change, so a further version bump
is required.

Signed-off-by: Eliad Peller <eliad.peller@wiz.io>